Refactoring to include default logger implementation.
