+++ src/automowerPlatform.ts
@@ -101,7 +101,7 @@ export class AutomowerPlatform implements DynamicPlatformPlugin {
             return;
         }        
 
-        this.container = new PlatformContainerImpl(this.config, this.api);
+        this.container = new PlatformContainerImpl(this.config, this.api, this.log);
         this.container.registerEverything();
     }
