+++ test/diagnostics/loggers/homebridgeImitationLogger.spec.ts
@@ -1,11 +1,11 @@
 import { It, Mock, Times } from 'moq.ts';
 
-import * as settings from '../../src/settings';
+import * as settings from '../../../src/settings';
 
-import { HomebridgeImitationLogger } from '../../src/diagnostics/platformLogger';
-import { ConsoleWrapper } from '../../src/primitives/consoleWrapper';
-import { Environment } from '../../src/primitives/environment';
-import { Localization } from '../../src/primitives/localization';
+import { HomebridgeImitationLogger } from '../../../src/diagnostics/loggers/homebridgeImitationLogger';
+import { ConsoleWrapper } from '../../../src/primitives/consoleWrapper';
+import { Environment } from '../../../src/primitives/environment';
+import { Localization } from '../../../src/primitives/localization';
 
 describe('HomebridgeImitationLogger', () => {
     let env: Mock<Environment>;
